tests/requirements-test.txt
pytest>=7.0.0
psycopg2-binary>=2.9.0
pyspark>=3.3.0